Rename : 패키지 구조 변경
UI 폴더 안에 액티비티와 다이얼로그, 어댑터, 프래그먼트를 모두 넣음
