+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
             android:value="5x22zex7t6" />
 
         <activity
-            android:name=".SplashActivity"
+            android:name=".ui.SplashActivity"
             android:exported="true"
             android:theme="@style/Theme.SeoulPublicService.RemoveSplash">
             <intent-filter>
@@ -34,10 +34,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".InterestRegionSelectActivity"
+            android:name=".ui.interestregionselect.InterestRegionSelectActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustNothing"
             android:label="@string/app_name">
