+++ app/src/main/java/com/wannabeinseoul/seoulpublicservice/ui/detail/DetailFragment.kt
@@ -25,11 +25,11 @@ import com.naver.maps.geometry.LatLng
 import com.naver.maps.map.*
 import com.naver.maps.map.overlay.Marker
 import com.naver.maps.map.util.MarkerIcons
-import com.wannabeinseoul.seoulpublicservice.MainViewModel
+import com.wannabeinseoul.seoulpublicservice.ui.main.MainViewModel
 import com.wannabeinseoul.seoulpublicservice.R
 import com.wannabeinseoul.seoulpublicservice.databases.ReservationEntity
 import com.wannabeinseoul.seoulpublicservice.databinding.FragmentDetailBinding
-import com.wannabeinseoul.seoulpublicservice.dialog.review.ReviewFragment
+import com.wannabeinseoul.seoulpublicservice.ui.dialog.review.ReviewFragment
 import com.wannabeinseoul.seoulpublicservice.util.loadWithHolder
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
