+++ app/src/main/java/com/wannabeinseoul/seoulpublicservice/usecase/CheckComplaintSelfUseCase.kt
@@ -1,7 +1,7 @@
 package com.wannabeinseoul.seoulpublicservice.usecase
 
 import com.wannabeinseoul.seoulpublicservice.databases.firebase.UserRepository
-import com.wannabeinseoul.seoulpublicservice.dialog.complaint.ComplaintUserInfo
+import com.wannabeinseoul.seoulpublicservice.ui.dialog.complaint.ComplaintUserInfo
 import com.wannabeinseoul.seoulpublicservice.pref.IdPrefRepository
 
 class CheckComplaintSelfUseCase(
