+++ app/src/main/java/com/wannabeinseoul/seoulpublicservice/usecase/ComplaintUserUseCase.kt
@@ -3,7 +3,7 @@ package com.wannabeinseoul.seoulpublicservice.usecase
 import com.wannabeinseoul.seoulpublicservice.databases.firebase.ComplaintEntity
 import com.wannabeinseoul.seoulpublicservice.databases.firebase.ComplaintRepository
 import com.wannabeinseoul.seoulpublicservice.databases.firebase.ReviewRepository
-import com.wannabeinseoul.seoulpublicservice.dialog.complaint.ComplaintUserInfo
+import com.wannabeinseoul.seoulpublicservice.ui.dialog.complaint.ComplaintUserInfo
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
