+++ app/src/main/java/com/wannabeinseoul/seoulpublicservice/usecase/GetReviewListUseCase.kt
@@ -2,7 +2,7 @@ package com.wannabeinseoul.seoulpublicservice.usecase
 
 import com.wannabeinseoul.seoulpublicservice.databases.firebase.ServiceRepository
 import com.wannabeinseoul.seoulpublicservice.databases.firebase.UserBanRepository
-import com.wannabeinseoul.seoulpublicservice.dialog.review.ReviewItem
+import com.wannabeinseoul.seoulpublicservice.ui.dialog.review.ReviewItem
 import com.wannabeinseoul.seoulpublicservice.pref.IdPrefRepository
 
 class GetReviewListUseCase(
