+++ app/src/main/res/layout/activity_interest_region_select.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".InterestRegionSelectActivity">
+    tools:context=".ui.interestregionselect.InterestRegionSelectActivity">
 
     <ImageView
         android:id="@+id/iv_interest_region_select_back_btn"
