+++ app/src/main/res/layout/fragment_complaint.xml
@@ -7,7 +7,7 @@
     android:background="@drawable/background_radius_12dp"
     android:elevation="5dp"
     android:clipToOutline="true"
-    tools:context=".dialog.complaint.ComplaintFragment">
+    tools:context=".ui.dialog.complaint.ComplaintFragment">
 
     <TextView
         android:id="@+id/tv_complaint_description"
