+++ app/src/main/res/layout/fragment_filter.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".dialog.filter.FilterFragment">
+    tools:context=".ui.dialog.filter.FilterFragment">
 
     <ImageView
         android:id="@+id/iv_filter_back_btn"
