+++ app/src/main/res/layout/fragment_review.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".dialog.review.ReviewFragment">
+    tools:context=".ui.dialog.review.ReviewFragment">
 
     <TextView
         android:id="@+id/tv_review_title"
